+++ tools/conv_miniboone.py
@@ -7,33 +7,33 @@
 import pandas as pd
 import numpy as np
 
-COL_LABEL = '_label'
+COL_LABEL = "_label"
 
 SEED = 998822
-#---
+# ---
 np.random.seed(SEED)
 
-data = pd.read_csv("../data/raw/miniboone.dat", header=None, sep=' +')
+data = pd.read_csv("../data/raw/miniboone.dat", header=None, sep=" +")
 
 data[COL_LABEL] = 0
 data.iloc[36500:][COL_LABEL] = 1
 
-data = data[ data[0] > -900]
+data = data[data[0] > -900]
 
-data.iloc[:, 0:-1] = data.iloc[:, 0:-1].astype('float32')
-data.iloc[:,-1:  ] = data.iloc[:,-1:  ].astype('int32')
+data.iloc[:, 0:-1] = data.iloc[:, 0:-1].astype("float32")
+data.iloc[:, -1:] = data.iloc[:, -1:].astype("int32")
 
 print(data.head())
 
 TRAIN_SIZE = 45359
-VAL_SIZE   = 19439
-TEST_SIZE  = 64798
+VAL_SIZE = 19439
+TEST_SIZE = 64798
 
 data = data.sample(frac=1)
 
 data_train = data.iloc[0:TRAIN_SIZE]
-data_val   = data.iloc[TRAIN_SIZE:TRAIN_SIZE+VAL_SIZE]
-data_test   = data.iloc[TRAIN_SIZE+VAL_SIZE:]
+data_val = data.iloc[TRAIN_SIZE : TRAIN_SIZE + VAL_SIZE]
+data_test = data.iloc[TRAIN_SIZE + VAL_SIZE :]
 
 print("Number of features:", data_train.shape[1] - 1)
 print("Classes:", data_train.iloc[:, -1].unique())
@@ -49,15 +49,15 @@ data_train.to_pickle("../data/mb-train")
 data_val.to_pickle("../data/mb-val")
 data_test.to_pickle("../data/mb-test")
 
-#--- prepare meta
+# --- prepare meta
 idx = data.columns[:-1]
-meta = pd.DataFrame(index=idx, dtype='float32')
+meta = pd.DataFrame(index=idx, dtype="float32")
 
-meta['avg'] = data_train.mean()
-meta['std'] = data_train.std()
-meta['cost'] = 1.
+meta["avg"] = data_train.mean()
+meta["std"] = data_train.std()
+meta["cost"] = 1.0
 
-meta = meta.astype('float32')
+meta = meta.astype("float32")
 
 print()
 print(meta)
